PiBasestation/BasestationToFirebaseIntegrator/requirements.txt
firebase-admin
pyserial
jsonschema
python-dotenv